meowth_fp: move large buffers
Configure the large arrays for the image frame and template in other
SRAM memory regions to free up space for the malloc heap in the 'main'
SRAM.

Note: for now, keep the image buffer where it is as it has the
convenient property of 'absorbing' the libalgorithm stack underflow.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=none
BUG=b:72360575
TEST=on ZerbleBarn, run the 'shmem' console and see 500kB of heap
available for malloc.

Change-Id: I31a1b4fb0d5e124e3229d29386cc6500e402dac0