gsctool: allow multiple sections for the same board
The same device could have several released firmware images, and also
several variations of the GBB contents. Any permutation of these two
sets is possible.

With this patch the verifier will continue looking for board section's
in the descriptors database even after a section was found.

Also added printout of processed ranges, to let the operator to see
the progress.

BRANCH=none
BUG=b:73668125
TEST=verified that both single and dual section descriptions are
     processed properly.

Change-Id: I181655192246ce245c43e6ce7ba4768ac5c9e51a
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>