Update patch set 2
Patch Set 2:

"Completely under our control" just means that only we have the ability to break it.  It doesn't mean we *can't* break it.

My concern is that we make a typo somewhere in the database.  

Say there's a new variant for Eve, with a different HWID and different RO firmware.  We add a new ZZAF section for the GBB.  And we add a new ZZAE section for the rest of RO firmware.  Yes, that's a typo.  And as usual, we're in a hurry so the new CL goes in at the last minute.

If we require exactly two matches, that will catch the typo.

If we have unit tests that test all variants of Eve firmware against the database, that will catch the typo.  (These wouldn't need to be done via CCD; we could compile a unit test gsctool with a mock interface so it just talks to data files.)

But if we only test one or two variants, and the new variant isn't one of them, then nobody will notice that we're not actually verifying the RO firmware for the new variant.

At least, nobody will notice until someone gets an open box Chromebook where the RO firmware has been modified, and something breaks.

Patch-set: 2
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>