Update patch set 2
Patch Set 2:

And actually, a typo the other direction is worse, because this verifies the entire GBB at once.

So, say we find a hash match in the RO section.

We don't find a GBB match, but since we had at least one match in at least one section, we call it good enough.

But that means the *entire* GBB isn't checked.  Including the GBB flags which force dev mode on and a short dev warning delay, and the root and recovery keys.

(That's one reason I suggested using the dump function to figure out where the HWID is in the GBB and hash everything but the HWID, and explicitly check HWIDs.  Though yes, that would require more explicit handling in gsctool.)

Patch-set: 2
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>