Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> We still need a part of this patch. We need to configure RPM values for Vayne. The aforementioned patch is just a preparation.

In the beginning, Nami and Vayne used different fan, so we created this CL.
This CL is for supporting two different configurations of fan in the same code.
To be more specific, Nami has one fan, and Vayne also has one fan, but the fan module are different between them, so we need to modify fan_rpm in run-time, after getted sku_id.
But now we changed the design, Nami and Vayne will use the same fan, so we no longer need this CL, we just need to modify the setting in fan_rpm_0. 
For this part, we will create other CL to combine all fan setting.

Patch-set: 5