driver: OPT3001: Add attenuation factor.
This commit adds the ability for a board to define an attenuation factor
that will be multiplied against the intermediate value prior to
converting to lux.  This helps for cases where lux readout may be very
low and higher precision is desired.

BUG=b:72985601
BRANCH=stabilize-meowth-10444.B
TEST=Flash meowth; verify reported readings are significantly higher
than before in low lux environments.

Change-Id: Ibf46e574fccdde50ceb5f8174f9b4a29e60bfbfe
Signed-off-by: Aseda Aboagye <aaboagye@google.com>