Support long waits with waitms
At present 'waitms 3000' trips the watchdog (at least on grunt). Add a
loop so we can support longer waits.

This uses 1024 (1 << 10) as an approximation for 1000, since this reduces
the code-size impact of this change. The error is only 2.5% and should not
matter for a CLI comand. On grunt:

Original text size:           265888
With decimal arithmetic:      265920
With shifts/masks (this CL):  265904

BUG=b:72542719
BRANCH=none
TEST=manually on grunt
> waitms 3000
See that it doesn't trip the watchdog now
(test various other times as well)

Change-Id: I453708299e4e26c1bbdb5fc406f26e916e7389af
Signed-off-by: Simon Glass <sjg@chromium.org>