Update patch set 1
Patch Set 1: Code-Review+1

I'm good with this, but as Ed mentioned in the bug, FAFT currently uses the waitms command to check the watchdog, so this change will cause that to fail.  If everyone else is ok with that, I certainly am.

The alternative I saw was to document that long delays will trip the watchdog in the help, and reset the watchdog when we leave the waitms function so that we can at least run multiple waitms 1000 commands in a row to have a long delay.  Currently that still resets.

Patch-set: 1
Reviewer: Gerrit User 1133034 <1133034@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1