base_detect_lux: Only allow enabling power if base is connected
Since we only enable power to the base in the charger task (we
can't enable power before OTG/power allocation is setup properly),
there is a potential race when we quickly plug/unplug the base.

Fix this by only allowing the enable power when the detection
code knows that the base is actually connected.

BRANCH=none
BUG=b:74403503
TEST=Plug/unplug wand several times with charger attached,
     PPVAR_VAR_BASE is always 0 whenever the base is disconnected.

Change-Id: I96da9ed39ddbbbe7aea54ad96c62fde6115538c0
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>