charge_state_v2: Simplify low-battery base power logic
When the base battery is low, we need to provide it with some power,
to make sure the battery does not fully deplete.

Instead of relying on a separate configuration value, we simply
reuse min_base_system_power. Also, we fix a bug that incorrectly
used a power value as current.

BRANCH=none
BUG=b:71881017
BUG=b:74414928
TEST=With a critically low wand (3%), lux is able to provide
     enough power to charge wand battery a little (44mA), even
     when backlight is fully on, and touchpad is in use.

Change-Id: I3174ac273712eeb83ce1283dda3d786ad503e6f3
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>