nds32: task: allow context switching if task_start() is called
We got a symptom that keyboard didn't work without connecting
servo board after this change (CL:897315) was merged.
This is because our uart RX will receive a data (0) with framing
error if RX level is low and trigger a re-scheduling request in
uart ISR (HOOKS task will be wake at later and then start the
task scheduling).
And that will cause we don't return to main() function to finish
all operations of initialization after uart_init() is called.

I think we will get the same symptom if GPIO/peripheral interrupts
are enabled and wake some task at initialization.

This change makes sure we will start the task scheduling if
task_start() is called.

BUG=none
BRANCH=none
TEST=With this change, keyboard works after EC reboot without
     servo board connected.

Change-Id: I0bda84b1cb56ced6aad2a38b0786d1b336e77211
Signed-off-by: Dino Li <Dino.Li@ite.com.tw>