link, pwm: enable fan when setting fan rpm via host command
Currently the flow setting the fan rpm via ectool command never enables
the fan, but expects it to be enabled. The ec console command does not
make that assumption, neither does the fanduty command (console, or host
command).
This always enables the fan in the ectool pwmsetfanrpm flow.

Note: This is the change required, but I had to make quite a few other
changes to get the ec image to even build. Unclear if that would cause
issues if we decided to pull this in.

BRANCH=link
BUG=chromium:808764
TEST=with this change
<on ec console>
>faninfo
...
Enabled: No
<on cpu console>
ectool pwmsetfanrpm 8000
<on ec console>
>faninfo
...
Enabled: Yes
(also hear the fan)

Change-Id: Ie18437150a1c252de645cdd6f76585a4ddba3536
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>