temp_metrics: use fanduty 0 to enable fan
If the fan is never enabled, temp_metrics itself has no code-path to
enable the fan. This is fixes this by calling fanduty 0 in the beginning
of temp_metrics, since fanduty does explicitly enable the fan.

Note: This is a hack to avoid having to flash a new EC image. See
crrev.com/c/964037 for a more fundamental fix to the same issue.

BRANCH=link
BUG=chromium:808764
TEST=couldn't reproduce issue with this version of temp_metrics.

Change-Id: I8a9b258ba7b50cf5180497d318f8d94454dab434
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>