USB-PD: Reset previously active port
When EC reboots, a USB-C device won't be notified because the CC line
stays pulled down. See b:74409830 for why we need this behavior.

If a dual-role device is SRC/UFP before reboot, it sees a data role
conflict because it sees the CC line is pulled down throughout the
reboot cycle. When the EC starts up, it becomes SNK/UFP, which
conflicts with the monitors data role. Some monitors issue a hard
reset to resolve this situation.

This patch ensures the port is reset if it's previously in an active
contract regardless how the EC resets.

BUG=b:74409830
BRANCH=fizz
TEST=Verify Fizz boots on HP S240n

Change-Id: I42665526ce1ec1c5b34d3b6e9e7ae38f06e145ca
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>