Update patch set 1
Patch Set 1:

I am not sure we want to push this particular change and the one that clears VSYS_PRIORITY after VBAT > VSYS unless there is an identified case where this is helping boot up with dead battery. On Eve there are present concerns about 'no battery' and recovering from ship mode or hibernate due to low battery with VSYS_PRIORITY set. The problem is that since this defeats the input current limit, it can cause the OCP of the external charger to trip and result in reboot loops when there is no battery present. Rohm is recommending that we leave VSYS_PRIORITY to the default value of 0 because of these concerns.
> 
> Also, is there a new RO being planned for the Reef FW branch?

Patch-set: 1
Reviewer: Gerrit User 1123814 <1123814@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>