Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> I wonder if we should keep it around in some form. lpc_host_reset() is a poor name, but we may need an espi_host_reset() vs lpc_host_reset(), etc.

Yeah, I had left it in because I was not sure if/how we would want to use this. It seems like a nice thing to have especially in case of eSPI it doesn't cost us anything as it can be done completely over virtual wires. But, no user for it right now -- so we can always add it back when there is a need.

Patch-set: 1
Reviewer: Gerrit User 1003000 <1003000@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>