meowth: Add GPP_B14_STRAP pins for both SKUs.
GPP_B14_STRAP is being moved due to a change in EC SKU.  We're not
currently using this pin, but if we decide to in the future, we can set
up the appropriate one based upon reading the Chip ID register.

BUG=none
BRANCH=None
TEST=Build and flash on both ECs, verify that they both boot up
normally.

Change-Id: Iaa25d5d77939bf55d6dc3991eec89ad5d6e92abb
Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/978677
Commit-Ready: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>