Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> > Patch Set 1:
> > 
> > A few things that confuse me here:
> > 1) Why do some boards use CONFIG_CHARGER_LIMIT_POWER_THRESH_* and some not? Coral's use of them was removed here: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/753651
> > 2) Do we really need both LIKELY_PD_USBC_POWER_MW and CONFIG_CHARGER_LIMIT_POWER_THRESH_CHG_MW?
> > 3) charge_manager_get_power_limit_uw() gets compared to LIKELY_PD_USBC_POWER_MW/CONFIG_CHARGER_LIMIT_POWER_THRESH_CHG_MW in 2 different places in this one function. Can those 2 checks be merged together?
> 
> 
> I will try to answer the questions based on my experience (but I could have missed out the historical aspect of this):
> 1) CONFIG_CHARGER_LIMIT_POWER_THRES_* expects that the AP does throttling based on the indication from EC that we have a low battery state-of-charge and/or weak charger is connected. On x86 platforms, there is nothing that we do on the AP side to handle this situation and hence it was reverted on eve, poppy, coral.


There is some code in depthcharge to wait for the EC to clear the limit power flag, I think  the idea is that the system needs to be able to boot into depthcharge with the available initial power and then will wait there until the EC confirms it has negotiated to a higher voltage and can continue to boot.

What is missing to make it useful in a dead battery scenario is something earlier in the BIOS to try and actually limit the power consumption when the EC has this flag set...  As long as the system can actually boot that far with reduced power it could still be useful.


> 
> 2) I think the idea here was that if a platform is okay with running in a throttled mode with  CONFIG_CHARGER_LIMIT_POWER_THRESH_CHG_MW where CONFIG_CHARGER_LIMIT_POWER_THRESH_CHG_MW is less than LIKELY_PD_USBC_POWER_MW, then it might be okay to boot it up as well even if the charger cannot provide LIKELY_PD_USBC_POWER_MW. I don't see any board setting it less than LIKELY_PD_USBC_POWER_MW though.
> 
> 3) I agree that from the looks of it, this function can be simplified. But we should be careful about not breaking any older board.

Yeah this whole function logic could use cleaning up, Shawn filed a bug but we didn't really make progress yet: crbug.com/782890

Patch-set: 1
Reviewer: Gerrit User 1000423 <1000423@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>