Update patch set 3
Patch Set 3: -Code-Review

> Patch Set 3: Code-Review-2
> 
> Actually, no, sorry. I don't think the change I'm requesting is unreasonable. If somebody CR+2 I'll remove my vote, but I don't think self-approving is acceptable.

No, it's not unreasonable. I got your points. Thank you for sharing them. Here is why I still think we should use abbreviations.

I'm aware consistency is important in source code. Unfortunately, past CL descriptions can't be changed. So, we will need to break the consistency if we want to introduce a new abbreviation.

Whether chgstv2 is a good abbreviation or not is something we can discuss or vote for but I suggest we should not use only the past commit messages to pick one or the other.

> I like consistency and clarity:

I think we should give more space to what follows after the preamble (e.g. chgstv2:). If we use abbreviations as a preamble, we can squeeze one or two more words after that. It helps people guess what the CL is about from the title. I think it improves clarity in a more significant sense. I believe that's what Aseda is trying to do as well in his CL.

For example, I use CBI = cros board info. It's not clear for the people who have never seen it. So, in that sense it's not clear. But once you know it, it's very useful. It provides a lot more space for developers to describe his/her patch in a single line.

It's a pity that we have to break the consistency (charge_state_v2) but I think the benefit is larger in the long run. (Ok, now it's 0 v.s. 40 but we'll get 10, 20, ... ;-)

Patch-set: 3
Reviewer: Gerrit User 1106949 <1106949@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0