Update patch set 1
Patch Set 1:

what happens if the other image is corrupted for whatever reason. In this case the check in system_rollback_detected() would return True, but it would not have been the case, as the RO would not run the other image.

I think paying attention to the counter and maybe verifying both images before deciding what happened would be more reliable.

Need to check if the public key is even included in the RW, or maybe we can find it in RO.

Patch-set: 1
Reviewer: Gerrit User 1000039 <1000039@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>