Update patch set 1
Patch Set 1:

I checked that after the header is corrupted during login, the system rollback state does not change.

system_rollback_detected is used to make actual decisions while cr50 is running, so if there is an issue with how that is calculated then it should be changed. 

The sysinfo is mostly for debugging/printing the system information. It should be consistent with what cr50 thinks the system state is.

Patch-set: 1
Reviewer: Gerrit User 1136059 <1136059@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>