Update patch set 1
Patch Set 1:

> I agree we should determine once if the rollback happened, and then use that result until next Cr50 reboot.

We'd need to deal with saving this in long life scratch or some or somewhere else, because this information would only be known until cr50 enters deep sleep.

If we make a change to system_rollback_detected, that should be in a different cl. This change just changes where the sysinfo command gets its rollback information. It is more in line with how cr50 determines rollback state, so I think this is better than what we have.

Patch-set: 1
Reviewer: Gerrit User 1136059 <1136059@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>