Update patch set 1
Patch Set 1:

This is the right way to do it in the long run.

But as Nicholas points out, the boot key detection code in the EC is part of EC-RO, so we can't update it on already-shipped Eve systems.  On those systems, the user will need to Power+Refresh and then hold down the 'A' key until the OS boots.

That's no worse than hitting F12 repeatedly to bring up the boot selection menu on any other laptop.

Patch-set: 1
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>