ite: exclude keyboard interrupts if no keyboard task
When building the `tests` target for a specific board that uses the
ite EC chip, it will fail to find the lpc_kbc_obe_interrupt function.

Adding #define around code similar to npcx set up.

This is needed because the unibuild system uses the tests target when
compiling chromeos-ec.

BRANCH=none
BUG=b:77274422
TEST=Can now successfully `make BOARD=bip tests`

Change-Id: I971aebe1667eb90e8ccccbe6047c5aa959c76c4b
Signed-off-by: Jett Rink <jettrink@chromium.org>