cr50: add vendor CCD subcommand to report CCD information
It is important for the OS to be able to find out the state of CCD and
its current capabilities settings of the device.

This patch adds a subcommand which would report 2 32bit values, one -
the bitmap of the current state of the CCD capabilities, the second
one - a number describing actual CCD state.

BRANCH=cr50, cr50-mp
BUG=b:72718383
TEST=with the following patch applied verified that CCD info is
     properly reported. Also verified that other CCD subcommands still
     work as advertised.

Change-Id: I4a783e6817ed364b9e64522ebbe968d4a657a84c
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>