Update patch set 1
Patch Set 1:

The command is ok, but it doesn't completely solve the problem in the referenced bug.  That is, we should send the equivalent of 'ccd get':

1. Current CCD state
2. Which capabilities are enabled
3. Setting for each capability (Always, UnlessLocked, IfOpened, etc.)
4. Flags

That is, if you're in 'ccd-opened' state, then by definition all capabilities are enabled, so you can't tell which ones would be disabled if you went into the locked state.

Patch-set: 1
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>