Fizz: Don't reset partner port on boot
Cros TCPM has a feature to reset the previously active partner ports
on boot. This feature is used only in recovery mode for Fizz.

It causes Fizz to unable to display the recovery screen on Acer XR342CK.
Since Cros TCPM works without it in normal/developer mode, we're
disabling the feature.

BUG=b/77722209
BRANCH=none
TEST=Test Fizz with XR342CK and other USB-C monitors in recovery,
normal, and developer mode.

Change-Id: I6214556605a2db30a116523582b0b6300b76267d
Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>