poppy: Prevent base detection from falling into reverse detection trap
The problem here is that both normal detection range (120-300mV)
and reverse connection detection range (450-500mV) are on the same
side of the interrupt pin TTL level.

So if we accidentally sample the ADC at the wrong time (either on
connection, or after a side-band wake signal from base), we may
fall into a trap where we assume that the base is connected in
reverse, while the value may just be a transient.

We have seen this case in the field, due to side-band wake pulse
from base EC, but this could potentially happen on attach as well.

The code debounces the reverse detection signal a little longer
(this has no impact as the base is not functional anyway), so that
we only consider that the base is connected in reverse if the ADC
value is within the range twice in a row.

BRANCH=poppy
BUG=b:77828249
TEST=With HACK CL to emulate bug:
 - Press a few keys, hammer gets disconnected
 - Connect/disconnect hammer a few times, sometimes it does not
   get disconnected.
TEST=Base connection, normal and reverse, still works properly.

Change-Id: I2ccb911472dd591146e3b0e68400c8bd55368dba
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1010044
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
(cherry picked from commit cdc9d417612275cdcc3b62f47c0a2cc6155b8c41)