Update patch set 8
Patch Set 8:

> Patch Set 8: Verified+1
> 
> It's time to push this CL in. It is verified on real hardware.
> 
> I don't think there's much to change, at least for the basic support. So I can work on some other CLs.

Do you have someone in mind on the EC team who would be a good reviewer for your patches (and also act as a backup for you for EC stuff?)  They should probably be assigned as Reviewer here.  One of the people you added this series could review this and +2, but it seems like it makes sense to find a secondary person on the EC team to do it.

...or does someone on this list want to do part time EC work and spearhead EC reviews?

Patch-set: 8
CC: Gerrit User 1112825 <1112825@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Gerrit User 1148900 <1148900@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Reviewer: Gerrit User 1000107 <1000107@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>