yorp: Keep 5V rail high on jump to RW
The BQ24392 driver controls power on the PP5000_A rail so that BC1.2
detection can be performed when the AP is off. However, that rail is
also controlled by the chipset task, so CONFIG_POWER_PP5000_CONTROL
needs to be defined to keep 5V rail up when necessary.

Calls to power_5v_enable() must only be done when the build has a
chipset task (otherwise, that function is not defined).

BRANCH=none
BUG=b:77874283
TEST=Booted Yorp with software sync enabled

Signed-off-by: Justin TerAvest <teravest@chromium.org>
Change-Id: Ib75655944aa05e381da922da8e916dc5d4dd9f85