cheza: HACK: Add apreset/apshutdown command
Adds apreset command that turns off the system with SWITCHCAP_ON_L, then
turns it back on with PMIC_KPD_PWR_ODL.

It also does this after EC reboot, in order to simulate normal
chromebook behavior.

This also makes the EC listen to the AP_RST_REQ pin, causing a cold
reset on request (for now).

Since apshutdown was so close / builtin the code I was about to enable.
That's now available as well.

Note, this is a temporary hack until we get rev 1 boards and until we
convince the PMIC to do some of our bidding (via fancy coreboot drivers
that haven't been written yet). Feel free to revert this once that
happens.

TEST=reboot/reprogram EC, see AP reboot
TEST=type "apreset", see AP reboot
BUG=none
BRANCH=none

Change-Id: I90f2ab35dd12d9120aa41e98784961965cc15a45
Signed-off-by: Alexandru M Stan <amstan@chromium.org>