cheza: Enable AP_RST_REQ as a request from AP to reset itself
This makes the EC listen to the AP_RST_REQ pin, calling the
chipset_reset() in the hook context, which bypasses the power
lost trigger during the reboot. So far the chipset_reset()
implementation is to do a cold reset; will be revised to a warm
reset after the PMIC registers are reprogrammed.

BRANCH=none
BUG=b:74395451
TEST=make buildall -j
TEST=Ran 'reboot' in AP userspace and checked AP reboot.

Change-Id: I946cb029541ce018a8ed1ce25681d38998a7f4b6
Signed-off-by: Wai-Hong Tam <waihong@google.com>